Implemented first PoC for FunctionImports
