+++ janos-core/src/test/java/org/apache/olingo/odata2/janos/processor/core/DataSourceProcessorTest.java
@@ -29,9 +29,9 @@ import org.mockito.Mockito;
 /**
  *
  */
-public class ListsProcessorTest {
+public class DataSourceProcessorTest {
 
-  private ListsProcessor listsProcessor;
+  private DataSourceProcessor dataSourceProcessor;
   private DataSource mockedDataSource = Mockito.mock(DataSource.class);
   private ValueAccess mockedValueAccess = Mockito.mock(ValueAccess.class);
   private DataStoreFactory mockedDataStoreFactory = Mockito.mock(DataStoreFactory.class);
@@ -40,57 +40,57 @@ public class ListsProcessorTest {
   public void init() throws ODataException {
     DataSource dataSource = new AnnotationDataSource(Building.class.getPackage().getName(), mockedDataStoreFactory);
     ValueAccess valueAccess = new AnnotationValueAccess();
-    ListsProcessor lp = new ListsProcessor(dataSource, valueAccess);
+    DataSourceProcessor lp = new DataSourceProcessor(dataSource, valueAccess);
 
     Assert.assertNotNull(lp);
   }
 
-  public ListsProcessorTest() {
-    listsProcessor = new ListsProcessor(mockedDataSource, mockedValueAccess);
+  public DataSourceProcessorTest() {
+    dataSourceProcessor = new DataSourceProcessor(mockedDataSource, mockedValueAccess);
   }
 
   @Test
   public void testSkipAndSkiptoken() {
     String url1 = "Rooms?$orderby=Seats%20desc&$skiptoken=12&$skip=000000&$top=200";
-    String result = listsProcessor.percentEncodeNextLink(url1);
+    String result = dataSourceProcessor.percentEncodeNextLink(url1);
     Assert.assertEquals("Rooms?$orderby=Seats%20desc&$top=200", result);
 
     String url2 = "Rooms?$orderby=Seats%20desc&$skiptoken=213&$skip=99";
-    String result2 = listsProcessor.percentEncodeNextLink(url2);
+    String result2 = dataSourceProcessor.percentEncodeNextLink(url2);
     Assert.assertEquals("Rooms?$orderby=Seats%20desc", result2);
 
     String url3 = "Rooms?$skiptoken=213&$skip=0000";
-    String result3 = listsProcessor.percentEncodeNextLink(url3);
+    String result3 = dataSourceProcessor.percentEncodeNextLink(url3);
     Assert.assertEquals("Rooms", result3);
   }
 
   @Test
   public void testSkipOnly() {
     String url = "Rooms?$orderby=Seats%20desc&$skip=000000&$top=200";
-    String result = listsProcessor.percentEncodeNextLink(url);
+    String result = dataSourceProcessor.percentEncodeNextLink(url);
     Assert.assertEquals("Rooms?$orderby=Seats%20desc&$top=200", result);
 
     String url2 = "Rooms?$orderby=Seats%20desc&$skip=213";
-    String result2 = listsProcessor.percentEncodeNextLink(url2);
+    String result2 = dataSourceProcessor.percentEncodeNextLink(url2);
     Assert.assertEquals("Rooms?$orderby=Seats%20desc", result2);
 
     String url3 = "Rooms?$skip=0999";
-    String result3 = listsProcessor.percentEncodeNextLink(url3);
+    String result3 = dataSourceProcessor.percentEncodeNextLink(url3);
     Assert.assertEquals("Rooms", result3);
   }
 
   @Test
   public void testSkiptokenOnly() {
     String url = "Rooms?$orderby=Seats%20desc&$skiptoken=213&$top=200";
-    String result = listsProcessor.percentEncodeNextLink(url);
+    String result = dataSourceProcessor.percentEncodeNextLink(url);
     Assert.assertEquals("Rooms?$orderby=Seats%20desc&$top=200", result);
 
     String url2 = "Rooms?$orderby=Seats%20desc&$skiptoken=213";
-    String result2 = listsProcessor.percentEncodeNextLink(url2);
+    String result2 = dataSourceProcessor.percentEncodeNextLink(url2);
     Assert.assertEquals("Rooms?$orderby=Seats%20desc", result2);
 
     String url3 = "Rooms?$skiptoken=213";
-    String result3 = listsProcessor.percentEncodeNextLink(url3);
+    String result3 = dataSourceProcessor.percentEncodeNextLink(url3);
     Assert.assertEquals("Rooms", result3);
   }
 }
