+++ janos-ref/src/main/java/org/apache/olingo/odata2/janos/processor/ref/model/RefFunctions.java
@@ -3,6 +3,8 @@ package org.apache.olingo.odata2.janos.processor.ref.model;
 import org.apache.olingo.odata2.api.annotation.edm.EdmFunctionImport;
 import org.apache.olingo.odata2.api.annotation.edm.EdmFunctionImportParameter;
 import org.apache.olingo.odata2.api.annotation.edm.EdmType;
+import org.apache.olingo.odata2.janos.processor.api.datasource.DataStoreFactory;
+import org.apache.olingo.odata2.janos.processor.api.datasource.FunctionExecutor;
 
 import static org.apache.olingo.odata2.api.annotation.edm.EdmFunctionImport.*;
 
@@ -11,11 +13,18 @@ import static org.apache.olingo.odata2.api.annotation.edm.EdmFunctionImport.*;
  *
  *
  */
-public class RefFunctions {
+public class RefFunctions implements FunctionExecutor {
+
+  private DataStoreFactory dataStoreFactory;
+
+  @Override
+  public void init(DataStoreFactory dataStore) {
+    dataStoreFactory = dataStore;
+  }
 
   @EdmFunctionImport(returnType = @ReturnType(type = ReturnType.Type.SIMPLE))
   public String citySearch(@EdmFunctionImportParameter(name = "cityName", type = EdmType.STRING) String name) {
-    // System.out.println("Search city: " + name);
+//    dataStoreFactory.
     return name;
   }
 }
